feat(story): 홈을 전광판으로 교체하고 일정을 /schedule로 분리
/ 에 무엇을 그릴지 고르는 스위치를 (main)/page.tsx에 두고, 전광판과
달력을 각자의 주소(/story·/schedule)로 살려 둔다. 리다이렉트가 아니라
고른 페이지를 그대로 그리므로 요청은 1회고 주소창도 그대로다. 롤백은
HOME_PAGE 상수만 되돌리면 끝이라 파일을 옮기지 않았다.

- 탭바: 가운데 로고 탭이 / , 일정 탭이 /schedule을 가리킨다
  (두 탭이 같은 화면을 가리키면 하나가 죽은 탭이 되므로 함께 손봤다)
- 인증 allowlist에 /schedule 추가 — 홈이 어느 쪽을 그리든 셋 다 공개 지면이다
- analytics tabClick: "home" -> "schedule". 과거 데이터의 "home"은
  달력 탭을 뜻하므로 지표 비교 시 주의
