+++ app/(info)/dev/story-styles/page.tsx
@@ -147,7 +147,7 @@ export default function StoryStylesPage() {
           지면 톤 9종은 모두 <strong>같은 10개 존·같은 순서</strong>로 그려집니다(톤만 비교).
         </p>
         <Link
-          href="/story"
+          href="/"
           className="rounded-full border border-border px-4 py-2 text-[12px] font-semibold text-foreground transition-colors hover:bg-muted focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
         >
           기강이야기로 돌아가기
